+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         </activity>
         <activity
 
-            android:name=".Diec.DiceActivity"
+            android:name=".Dice.DiceActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
